don't think this is the right approach
problem is the last layer is like a key -- it's obvious that only a
certain activation pattern is going to be valid to get through the
entire model.

how do we find it? it's not easy. I could try to hardcode the obvious
patterns (single vector firing etc) but this gets combinatorily more
complex, I assume they fire a good few of those vectors like a key.
hmmmm